requirements.txt
amqp==5.0.1
appdirs==1.4.4
asgiref==3.2.8
asn1crypto==0.24.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
boto3==1.13.24
botocore==1.16.24
cairocffi==1.2.0
CairoSVG==2.4.2
celery==5.0.0
certifi==2020.4.5.1
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==2.1.4
cssselect2==0.3.0
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.2
django-celery-beat==2.1.0
django-cors-headers==3.2.1
django-crontab==0.7.1
django-extensions==3.0.9
django-filter==2.2.0
django-mysql==3.8.1
django-sslserver==0.22
django-timezone-field==4.0
django-redis-sessions==0.6.2
lxml
djangorestframework==3.11.0

docutils==0.15.2
ecdsa==0.15
et-xmlfile==1.0.1
filelock==3.0.12
html5lib==1.1
httmock==1.3.0
identify==1.4.25
idna==2.9
importlib-metadata==2.0.0
jdcal==1.4.1
jmespath==0.10.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.0.2
mypy-extensions==0.4.3
mysqlclient==1.4.6
nodeenv==1.4.0
numpy==1.18.5
openpyxl==3.0.3
pandas==1.0.4
pathspec==0.8.0
Pillow==8.0.0
pre-commit==2.6.0
prompt-toolkit==3.0.7
pyasn1==0.4.8
pycparser==2.20
pycrypto==2.6.1
pygal==2.4.0

Pyphen==0.9.5
python-crontab==2.5.1
python-dateutil==2.8.1
python-jose==3.1.0
pytz==2020.1
pyxdg==0.25
PyYAML==5.3.1
redis==3.5.3
regex==2020.10.23
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
SecretStorage==2.3.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
tinycss2==1.0.2
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
vine==5.0.0
virtualenv==20.0.30
wcwidth==0.2.5
WeasyPrint==51
webencodings==0.5.1
xlrd==1.2.0
zipp==3.3.0
django-storages
woocommerce==2.1.1
ipython

ujson~=4.0.2
drf-yasg==1.20.0
